chore: complete signals migration and cleanup old form files
- Remove old useState-based form implementations (ArtistForm, CategoryForm, ProductForm)
- Rename *Signals files to standard names (ArtistFormSignals → ArtistForm, etc.)
- Update all imports and exports to use renamed components
- Add preventAutoPublish checkbox to ArtistForm
- Update documentation to reflect migration completion:
  - AGENTS.md: Signals status → Complete
  - ADR-015: Status → Implemented
  - SIGNALS-ADOPTION-PLAN.md: Mark all checklist items complete
  - SIGNALS-MIGRATION-GUIDE.md: Add completion note

The signals-based forms are now the standard implementation.
Old useState versions have been removed.
