+++ apps/maple-spruce/src/components/artists/index.ts
@@ -1,3 +1,2 @@
 export { ArtistForm } from './ArtistForm';
-export { ArtistFormSignals } from './ArtistFormSignals';
 export { ArtistList } from './ArtistList';
