+++ apps/maple-spruce/src/components/categories/index.ts
@@ -1,3 +1,2 @@
 export { CategoryForm } from './CategoryForm';
-export { CategoryFormSignals } from './CategoryFormSignals';
 export { CategoryList } from './CategoryList';
