+++ apps/maple-spruce/src/components/inventory/index.ts
@@ -1,5 +1,4 @@
 export { ProductForm } from './ProductForm';
-export { ProductFormSignals } from './ProductFormSignals';
 export { ProductList } from './ProductList';
 export { ProductDataTable } from './ProductDataTable';
 export { ProductFilterToolbar, defaultFilters } from './ProductFilterToolbar';
